functions/azure-function-cosmos-synapse-link/FunctionApp/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
azure-core==1.8.2         # via azure-cosmos
azure-cosmos==4.1.0       # via -r requirements.in
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
idna==2.10                # via requests
requests==2.24.0          # via azure-core
six==1.15.0               # via azure-core, azure-cosmos
urllib3==1.25.10          # via requests
